Animation and torque and slipstream fix by Martin Fenelon
When two or more of these aircraft are together on MP, animations on my
aircraft, canopy opening for example, are also seen on all others. This
patch fixes this behaviour.

The torque and slipstream effects appear to be in the wrong sense.
Application of power should cause a roll and yaw to the left with that
particular prop. At present this is reversed.